Tests: Also run common tests in custom test suite.
